pyproject.toml
[project]
name = "z3-embedding-quality"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.7",
    "numpy>=2.3.4",
    "pandas>=2.3.3",
    "scikit-learn>=1.7.2",
    "scipy>=1.16.3",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
    "z3-solver>=4.15.4.0",
]
